separated network service
